make navbar component and move links off main app file into it
